imagemagick based function to transform the RGB color space into another

Imagemagick is used for the calculation of the transformations. Please provide a tif file

colorspace(
  input = NULL,
  colorspace = c("cielab", "CMY", "Gray", "HCL", "HSB", "HSI", "Log", "XYZ", "YUV"),
  compress = "None",
  depth = 8,
  verbose = FALSE,
  retRaster = TRUE
)

Arguments

input

character. name of a (Geo)Tiff containing RGB data channels

colorspace

character. For a list of argument to determine colorspace see colorspace manual

compress

character. compression type depending on the imagemagick version the choices are: None, BZip, Fax, Group4, JPEG, JPEG2000, Lossless, LZW, RLE or Zip.

depth

numeric. color space depth in bit default is 8

verbose

be quiet

retRaster

logical if true a rasterstack is returned

Value

raster* object

Note

you need to install imagemagick on your system, please look at see imagemagick download section.

Examples