merge single images

merge single geocoded images to on big image

merge_rgb(
  files = NULL,
  proj4 = NULL,
  output = NULL,
  cropoutput = NULL,
  ext = NULL
)

Arguments

files

list of filenames to be merged

proj4

valid proj4 string

output

valid path and name for the output image

ext

valid extent of the the area that is requested

Value

The merged image as a raster object # example # tif_files is a list of valid filenames merged_r = merge_rgb(files = tif_files, out = file.path(tempdir(),"out.tif"))

Details

https://stackoverflow.com/questions/15876591/merging-multiple-rasters-in-r with regards to Flo Detsch who asked it the right way

Note

if proj4 is NULL no reprojection is performed otherwise the merged raster is projected by using the proj4 string.